requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
httpx==0.28.1
botbuilder-integration-aiohttp==4.17.1
jinja2==3.1.6
python-multipart==0.0.20
